% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_ph2bye.R
\docType{package}
\name{ph2bye-package}
\alias{ph2bye}
\alias{ph2bye-package}
\title{ph2bye: A package for Phase II single-arm Bayesian design.}
\description{
The ph2bye package provides three categories of important functions:
PostP.design, PredP.design and MultPostP.design.
}
\section{Posterior probability criterion functions}{

The posterior probability criterion functions include PostP and PostP.design functions.
}

\section{Predictive probability criterion functions}{

The predictive probability criterion functions include PredP and PredP.design functions.
}

\section{Posterior probability criterion function for multiple outcomes}{

The posterior probability criterion functions include MultPostP and MultPostP.design functions.
}

\section{Whole design function with double thresholds showing futility and efficacy boundary together}{

The criterion function DT.design.
}

\section{Prior calculation function}{

The function prior calculating Beta prior parameters accourding different prior information.
}
\author{
Yalin Zhu <yalin.zhu@outlook.com>
}

